% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D10_data.R
\docType{data}
\name{D10_data}
\alias{D10_data}
\title{D10 Artificial Dataset}
\format{
A data frame with 1250 data points and 3 variables
\describe{
  \item{\code{x}}{Numeric values generated from Gaussian and Uniform distributions}
  \item{\code{y}}{Numeric values generated from Gaussian and Uniform distributions}
  \item{\code{label}}{Categorical labels {1,2,3,4,5}}
}
}
\usage{
D10_data
}
\description{
A \code{2}-dimensional dataset from Wiroonsri and Preedasawakul (2023) generated from \code{3} different Gaussian and \code{2} Uniform distributions labeled as \code{1-5}.
}
\references{
{N. Wiroonsri, O. Preedasawakul, A correlation-based fuzzy cluster validity index with secondary options detector, arXiv:2308.14785, 2023}
}
\seealso{
{\link{FzzyCVIs},  \link{WP.IDX},  \link{D1_data}, \link{Hvalid}, \link{DI.IDX}}
}
\author{
{Nathakhun Wiroonsri and Onthada Preedasawakul}
}
\keyword{datasets}
