% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1_data.R
\docType{data}
\name{D1_data}
\alias{D1_data}
\title{D1 Artificial Dataset}
\format{
A data frame with 1500 data points and 3 variables
\describe{
  \item{\code{x}}{Numeric values generated from Gaussian distributions}
  \item{\code{y}}{Numeric values generated from Gaussian distributions}
  \item{\code{label}}{Categorical labels {1,2,3,4,5,6}}
}
}
\usage{
D1_data
}
\description{
A \code{2}-dimensional dataset from Wiroonsri and Preedasawakul (2023) generated from \code{6} different Gaussian distributions labeled as \code{1-6}.
}
\references{
{N. Wiroonsri, O. Preedasawakul, A correlation-based fuzzy cluster validity index with secondary options detector, arXiv:2308.14785, 2023}
}
\seealso{
{\link{FzzyCVIs},  \link{WP.IDX},  \link{D1_data}, \link{Hvalid}, \link{DI.IDX}}
}
\author{
{Nathakhun Wiroonsri and Onthada Preedasawakul}
}
\keyword{datasets}
