% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D5_data.R
\docType{data}
\name{D5_data}
\alias{D5_data}
\title{D5 Artificial Dataset}
\format{
A \code{data frame} with 350 data points and 3 variables
\describe{
  \item{\code{x}}{Numeric values generated from Gaussian distributions}
  \item{\code{y}}{Numeric values generated from Gaussian distributions}
  \item{\code{label}}{Categorical labels 1,2,3,4,5}
}
}
\usage{
D5_data
}
\description{
A \code{2}-dimensional dataset from Wiroonsri and Preedasawakul (2023) generated from \code{5} different Gaussian distributions labeled as \code{1-5}.
}
\references{
{N. Wiroonsri, O. Preedasawakul, A correlation-based fuzzy cluster validity index with secondary options detector, arXiv:2308.14785, 2023}
}
\seealso{
{\link{FzzyCVIs},  \link{WP.IDX},  \link{D1_data}, \link{Hvalid}, \link{DI.IDX}}
}
\author{
{Nathakhun Wiroonsri and Onthada Preedasawakul}
}
\keyword{datasets}
