% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D9_data.R
\docType{data}
\name{D9_data}
\alias{D9_data}
\title{D9 Artificial Dataset}
\format{
A data frame with 1000 data points and 3 variables
\describe{
  \item{\code{x}}{Numeric values generated from Uniform distributions}
  \item{\code{y}}{Numeric values generated from Uniform distributions}
  \item{\code{label}}{Categorical labels {1,2,3}}
}
}
\usage{
D9_data
}
\description{
A \code{2}-dimensional dataset from Wiroonsri and Preedasawakul (2023) generated from \code{3} different Uniform distributions labeled as \code{1-3}.
}
\references{
{N. Wiroonsri, O. Preedasawakul, A correlation-based fuzzy cluster validity index with secondary options detector, arXiv:2308.14785, 2023}
}
\seealso{
{\link{FzzyCVIs},  \link{WP.IDX},  \link{D1_data}, \link{Hvalid}, \link{DI.IDX}}
}
\author{
{Nathakhun Wiroonsri and Onthada Preedasawakul}
}
\keyword{datasets}
